requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.11.0
openai==2.3.0
psycopg2-binary==2.9.9
pydantic==2.12.2
pydantic_core==2.41.4
python-dotenv==1.0.0
PyYAML==6.0.3
redis==6.4.0
regex==2025.9.18
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
tiktoken==0.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.1
websockets==15.0.1
